requirements.txt
fastapi==0.115.12
langchain_openai==0.3.19
langchain_tableau==0.4.39
langsmith==0.3.44
pydantic==2.11.5
python-dotenv==1.1.0
uvicorn==0.34.3
tableauserverclient
